Fixes critical bug regarding HashMap_Node copying.
It turns out due to its destructor, a naive copy of a HashMap_Node has
unacceptable side effects. The HashMap_Node manages dynamically-allocated
memory, so a proper copy constructor would require allocation of a Key and a
Value. Since Values are expected to be able to be arbitrarily large and since
copying is completely avoidable, we will just not do it.